pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "private-exposure"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.1",
  "sqlmodel>=0.0.38",
  "psycopg2-binary>=2.9.11",
  "fastapi>=0.115.0",
  "uvicorn>=0.34.0",
  "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: tests that call live external services",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]